baby_generator_tri/requirements.txt
dlib @ https://github.com/davisking/dlib
numpy
cv2
keras
sklearn
pickle
skimage
Pillow